feat(frame): 闪屏重构，支持动画效果
