+++ frame/src/main/resources/com/tlcsdm/frame/i18n/messages_en.properties
@@ -35,4 +35,3 @@ frame.sample.rightPanel.sampleVersion=Component Version
 frame.main.confirmExit.title=Exit
 frame.main.confirmExit.message=Are you sure you want to quit?
 frame.splash.init.system=Initialization system configuration...
-frame.splash.init.version=Version detection...
