+++ frame/src/main/resources/com/tlcsdm/frame/i18n/messages_ja.properties
@@ -35,4 +35,3 @@ frame.sample.rightPanel.sampleVersion=\u30B3\u30F3\u30DD\u30FC\u30CD\u30F3\u30C8
 frame.main.confirmExit.title=\u51FA\u53E3
 frame.main.confirmExit.message=\u672C\u5F53\u306B\u3084\u3081\u308B\u6C17\uFF1F
 frame.splash.init.system=\u521D\u671F\u5316\u30B7\u30B9\u30C6\u30E0\u69CB\u6210...
-frame.splash.init.version=\u30D0\u30FC\u30B8\u30E7\u30F3\u306E\u691C\u51FA...
