+++ frame/src/main/resources/com/tlcsdm/frame/i18n/messages_zh.properties
@@ -35,4 +35,3 @@ frame.sample.rightPanel.sampleVersion=\u7EC4\u4EF6\u7248\u672C\u53F7
 frame.main.confirmExit.title=\u9000\u51FA\u5E94\u7528
 frame.main.confirmExit.message=\u786E\u5B9A\u8981\u9000\u51FA\u5417\uFF1F
 frame.splash.init.system=\u521D\u59CB\u5316\u7CFB\u7EDF\u914D\u7F6E...
-frame.splash.init.version=\u7248\u672C\u68C0\u6D4B...
